add config of user precompiled contract && add support for docker

+++ tools/build_chain.sh
@@ -23,6 +23,7 @@ logfile=build.log
 listen_ip="127.0.0.1"
 bcos_bin_name=fisco-bcos
 guomi_mode=
+docker_mode=
 gm_conf_path="gmconf/"
 current_dir=$(pwd)
 consensus_type="pbft"
@@ -69,7 +70,7 @@ LOG_INFO()
 
 parse_params()
 {
-while getopts "f:l:o:p:e:t:icszhgTF" option;do
+while getopts "f:l:o:p:e:t:icszhgTFd" option;do
     case $option in
     f) ip_file=$OPTARG
        use_ip_param="false"
@@ -92,6 +93,7 @@ while getopts "f:l:o:p:e:t:icszhgTF" option;do
     F) auto_flush="false";;
     z) make_tar="yes";;
     g) guomi_mode="yes";;
+    d) docker_mode="yes";;
     h) help;;
     esac
 done
@@ -675,25 +677,38 @@ EOF
 generate_node_scripts()
 {
     local output=$1
+    local docker_tag="latest"
     generate_script_template "$output/start.sh"
+    local ps_cmd="\`ps aux|grep \${fisco_bcos}|grep -v grep|awk '{print \$2}'\`"
+    local start_cmd="nohup \${fisco_bcos} -c config.ini 2>>nohup.out"
+    local stop_cmd="kill \${node_pid}"
+    local pid="pid"
+    local log_cmd="cat nohup.out"
+    if [ ! -z ${docker_mode} ];then
+        ps_cmd="\`docker ps |grep \${SHELL_FOLDER//\//} | grep -v grep|awk '{print \$1}'\`"
+        start_cmd="docker run -d --rm --name \${SHELL_FOLDER//\//} -v \${SHELL_FOLDER}:/data --network=host -w=/data fiscoorg/fiscobcos:${docker_tag} -c config.ini >>nohup.out"
+        stop_cmd="docker kill \${node_pid} 2>/dev/null"
+        pid="container id"
+        log_cmd="docker logs \${SHELL_FOLDER//\//}"
+    fi
     cat << EOF >> "$output/start.sh"
 fisco_bcos=\${SHELL_FOLDER}/../${bcos_bin_name}
 cd \${SHELL_FOLDER}
 node=\$(basename \${SHELL_FOLDER})
-node_pid=\`ps aux|grep "\${fisco_bcos}"|grep -v grep|awk '{print \$2}'\`
+node_pid=${ps_cmd}
 if [ ! -z \${node_pid} ];then
-    echo " \${node} is running, pid is \$node_pid."
+    echo " \${node} is running, ${pid} is \$node_pid."
     exit 0
 else 
-    nohup \${fisco_bcos} -c config.ini 2>>nohup.out &
-    sleep 0.5
+    ${start_cmd} &
+    sleep 1
 fi
-node_pid=\`ps aux|grep "\${fisco_bcos}"|grep -v grep|awk '{print \$2}'\`
+node_pid=${ps_cmd}
 if [ ! -z \${node_pid} ];then
     echo -e "\033[32m \${node} start successfully\033[0m"
 else
     echo -e "\033[31m \${node} start failed\033[0m"
-    cat nohup.out
+    ${log_cmd}
     exit 1
 fi
 EOF
@@ -701,7 +716,7 @@ EOF
     cat << EOF >> "$output/stop.sh"
 fisco_bcos=\${SHELL_FOLDER}/../${bcos_bin_name}
 node=\$(basename \${SHELL_FOLDER})
-node_pid=\`ps aux|grep "\${fisco_bcos}"|grep -v grep|awk '{print \$2}'\`
+node_pid=${ps_cmd}
 try_times=5
 i=0
 while [ \$i -lt \${try_times} ]
@@ -710,11 +725,11 @@ do
         echo " \${node} isn't running."
         exit 0
     fi
-    [ ! -z \${node_pid} ] && kill \${node_pid}
-    sleep 0.5
-    node_pid=\`ps aux|grep "\${fisco_bcos}"|grep -v grep|awk '{print \$2}'\`
+    [ ! -z \${node_pid} ] && ${stop_cmd} 2> /dev/null
+    sleep 0.6
+    node_pid=${ps_cmd}
     if [ -z \${node_pid} ];then
-        echo " stop \${node} success."
+        echo -e "\033[32m stop \${node} success.\033[0m"
         exit 0
     fi
     ((i=i+1))
@@ -812,9 +827,9 @@ parse_ip_config()
     local config=$1
     n=0
     while read line;do
-        ip_array[n]=$(echo ${line} | cut -d ' ' -f 1)
-        agency_array[n]=$(echo ${line} | cut -d ' ' -f 2)
-        group_array[n]=$(echo ${line} | cut -d ' ' -f 3)
+        ip_array[n]=$(echo ${line} | awk '{print $1}')
+        agency_array[n]=$(echo ${line} | awk '{print $2}')
+        group_array[n]=$(echo ${line} | awk '{print $3}')
         if [ -z "${ip_array[$n]}" -o -z "${agency_array[$n]}" -o -z "${group_array[$n]}" ];then
             LOG_WARN "Please check ${config}, make sure there is no empty line!"
             return 1
@@ -843,32 +858,34 @@ dir_must_not_exists ${output_dir}
 mkdir -p "${output_dir}"
 
 # download fisco-bcos and check it
-if [ -z ${bin_path} ];then
-    bin_path=${output_dir}/${bcos_bin_name}
-    package_name="fisco-bcos.tar.gz"
-    [ ! -z "$guomi_mode" ] && package_name="fisco-bcos-gm.tar.gz"
-    version=$(curl -s https://raw.githubusercontent.com/FISCO-BCOS/FISCO-BCOS/master/release_note.txt | sed "s/^[vV]//")
-    Download_Link="https://github.com/FISCO-BCOS/FISCO-BCOS/releases/download/v${version}/${package_name}"
-    LOG_INFO "Downloading fisco-bcos binary from ${Download_Link} ..." 
-    curl -LO ${Download_Link}
-    tar -zxf ${package_name} && mv fisco-bcos ${bin_path} && rm ${package_name}
-    chmod a+x ${bin_path}
-else
-    echo "Checking fisco-bcos binary..."
-    bin_version=$(${bin_path} -v)
-    if [ -z "$(echo ${bin_version} | grep 'FISCO-BCOS')" ];then
-        LOG_WARN "${bin_path} is wrong. Please correct it and try again."
-        exit 1
+if [ -z ${docker_mode} ];then
+    if [ -z ${bin_path} ];then
+        bin_path=${output_dir}/${bcos_bin_name}
+        package_name="fisco-bcos.tar.gz"
+        [ ! -z "$guomi_mode" ] && package_name="fisco-bcos-gm.tar.gz"
+        version=$(curl -s https://raw.githubusercontent.com/FISCO-BCOS/FISCO-BCOS/master/release_note.txt | sed "s/^[vV]//")
+        Download_Link="https://github.com/FISCO-BCOS/FISCO-BCOS/releases/download/v${version}/${package_name}"
+        LOG_INFO "Downloading fisco-bcos binary from ${Download_Link} ..." 
+        curl -LO ${Download_Link}
+        tar -zxf ${package_name} && mv fisco-bcos ${bin_path} && rm ${package_name}
+        chmod a+x ${bin_path}
+    else
+        echo "Checking fisco-bcos binary..."
+        bin_version=$(${bin_path} -v)
+        if [ -z "$(echo ${bin_version} | grep 'FISCO-BCOS')" ];then
+            LOG_WARN "${bin_path} is wrong. Please correct it and try again."
+            exit 1
+        fi
+        if [[ ! -z ${guomi_mode} && -z $(echo ${bin_version} | grep 'gm') ]];then
+            LOG_WARN "${bin_path} isn't gm version. Please correct it and try again."
+            exit 1
+        fi
+        if [[ -z ${guomi_mode} && ! -z $(echo ${bin_version} | grep 'gm') ]];then
+            LOG_WARN "${bin_path} isn't standard version. Please correct it and try again."
+            exit 1
+        fi
+        echo "Binary check passed."
     fi
-    if [[ ! -z ${guomi_mode} && -z $(echo ${bin_version} | grep 'gm') ]];then
-        LOG_WARN "${bin_path} isn't gm version. Please correct it and try again."
-        exit 1
-    fi
-    if [[ -z ${guomi_mode} && ! -z $(echo ${bin_version} | grep 'gm') ]];then
-        LOG_WARN "${bin_path} isn't standard version. Please correct it and try again."
-        exit 1
-    fi
-    echo "Binary check passed."
 fi
 
 if [ -z ${CertConfig} ] || [ ! -e ${CertConfig} ];then
@@ -1059,7 +1076,7 @@ for line in ${ip_array[*]};do
         ip_node_counts[${ip//./}]=$(( ${ip_node_counts[${ip//./}]} + 1 ))
     done
     generate_server_scripts "${output_dir}/${ip}"
-    cp "$bin_path" "${output_dir}/${ip}/fisco-bcos"
+    if [ -z ${docker_mode} ];then cp "$bin_path" "${output_dir}/${ip}/fisco-bcos"; fi
     if [ -n "$make_tar" ];then cd ${output_dir} && tar zcf "${ip}.tar.gz" "${ip}" && cd ${current_dir};fi
     ((++server_count))
 done
